iterating/iterating.go
package main

func Sum(numbers []int) int {
	sum_arr :=0
	// range vraca dve vrednosti, prva je indeks a druga vrednost i da bismo uzeli vrednost
	// koristimo _ koji preskace vrednost na koju se odnosi
	for _, number := range numbers {
		sum_arr += number
	}
	return sum_arr
}

/*prvi nacin
func SumAll(numbersToSum ...[]int) []int {
	lengthOfNumbers := len(numbersToSum)
	sums := make([]int, lengthOfNumbers)

	for i, numbers := range numbersToSum {
		sums[i] = Sum(numbers)
	}
	return sums
}*/

//drugi nacin
func SumAll(numbersToSum ...[]int) []int {
	var sums []int
	for _, numbers := range numbersToSum {
		sums = append(sums, Sum(numbers))
	}
	return sums
}

func SumAllTails(numbersToSum ...[]int) []int {
	var sums []int
	for _, numbers := range numbersToSum {
		if len(numbers) == 0 {
			sums = append(sums, 0)
		} else {
			tail := numbers[1:]
			sums = append(sums, Sum(tail))
		}
	}
	return sums
}